Fully Fleshed out AnimeCharacter superclass
This superclass is meant to act as a hierarchical figure to the subclasses OnePiece and Jujutsu Kaisen.  It implements the Fighter and Comparable interfaces. From fighter, it generic abilites which all subclasses will also inherit, but in ways that affect their own uniqure instance variables. It also operates as a figure from which OnePiece and Jujutsukaisen will inherit behavior, such as attacking and combo attacking, as well as basic features that all subclasses will inherit like hp and attack power. it also has an equals, comparable, and tostring method which subclasses will call on using super. When using the default constructor, the character's stats are randomly created and a random character is retrievered from the arraylist instance variable called characters. When the custom constructor is called, if the character is not in the arraylist characters, a character is forcibly chosen for them.