JJKAbilities Finished, new method in AnimeCharacter
I added a randomzieCharacter method in the superclass, where if they wanted to randomly change their character name they could. It uses a for loop to iterate through the available names and stops on a random one. In my JJK abilities class, I have 2 private instance variables that are mainly used, CE and CT. The constructors call the super(), and set the rest of the instance variables. My useCT method uses a technique on the opponent, and depending on the technique they have, it does differing amounts of damage. if the opponent hp goes below 0, it is game over. my recover and gameOver methods in the subclass now affect cursedEnergy and cursedTechnique.