Finished my OnePiece Abilities Class
Created a OnepieceABilities class which grants abilites from onepiece to anime characters. it extends to AnimeCharacter, and uses super when calling in the constructor and methods. my methods are unique here as I determine the amount of damage I can do depending on whteher I have Haki, a devil fruit, or both. It does increasing damage in order of devil fruit, Haki, and both, but has a higher risk of messing up because of energy requirements. my recover method in this class helps with it as it recovers my haki and devil fruit powers to their initial state, and my game over method also sets everything to 0 to signify death.